Add global exception handler

+++ src/main/java/com/skillmentor/core/controller/MentorController.java
@@ -1,27 +1,22 @@
 package com.skillmentor.core.controller;
 
 import com.skillmentor.core.dto.MentorDTO;
+import com.skillmentor.core.exception.MentorNotFoundException;
 import com.skillmentor.core.service.MentorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/mentor")
 public class MentorController {
+
     @Autowired
     private MentorService mentorService;
+
     @GetMapping
     public ResponseEntity<List<MentorDTO>> getAllMentor(
             @RequestParam(required = false) List<String> firstNames,
@@ -38,25 +33,31 @@ public class MentorController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<MentorDTO> findMentorById(@PathVariable Integer id){
-        final MentorDTO mentorDTO = mentorService.findMentorById(id);
-        return new ResponseEntity<>(mentorDTO, HttpStatus.OK);
+    public ResponseEntity<MentorDTO> findMentorById(@PathVariable Integer id) {
+        MentorDTO mentorDTO = mentorService.findMentorById(id);
+        if (mentorDTO == null) {
+            throw new MentorNotFoundException("Mentor with ID " + id + " not found");
+        }
+        return ResponseEntity.ok(mentorDTO);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<MentorDTO> deleteMentorById(@PathVariable Integer id){
-        final MentorDTO mentorDTO =  mentorService.deleteMentorById(id);
-        return new ResponseEntity<>(mentorDTO, HttpStatus.OK);
+    public ResponseEntity<MentorDTO> deleteMentorById(@PathVariable Integer id) {
+        if (id <1) {
+            throw new MentorNotFoundException("Cannot delete, Mentor with ID " + id + " not found");
+        }
+        MentorDTO mentorDTO = mentorService.deleteMentorById(id);
+        return ResponseEntity.ok(mentorDTO);
     }
 
     @PostMapping
-    public ResponseEntity<MentorDTO> createMentor(@RequestBody MentorDTO mentorDTO){
+    public ResponseEntity<MentorDTO> createMentor(@RequestBody MentorDTO mentorDTO) {
         final MentorDTO mentor = mentorService.createMentor(mentorDTO);
-        return new ResponseEntity<>(mentor, HttpStatus.OK);
+        return new ResponseEntity<>(mentor, HttpStatus.CREATED);
     }
 
     @PutMapping
-    public ResponseEntity<MentorDTO> updateMentor(@RequestBody MentorDTO mentorDTO){
+    public ResponseEntity<MentorDTO> updateMentor(@RequestBody MentorDTO mentorDTO) {
         final MentorDTO mentor = mentorService.updateMentor(mentorDTO);
         return new ResponseEntity<>(mentor, HttpStatus.OK);
     }
